@@ -5,14 +5,14 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.10.4 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 #  firebase_auth: ^0.18.3+1
-  firebase_auth: ^0.18.2
-  google_sign_in: ^4.5.6
+  firebase_auth: ^1.0.1
+  google_sign_in: ^5.0.0
 
   engine_db_utils:
     git:
