@@ -4,7 +4,6 @@ import 'package:engine_db_utils/models/log.dart';
 import 'package:engine_db_utils/models/result.dart';
 import 'package:engine_utils/utils/string_utils.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 
 const String errNull = "Got a null auth user from backend";
@@ -12,19 +11,19 @@ const String errNull = "Got a null auth user from backend";
 abstract class AuthService{
 
   Future<Result<AuthUser>> registerWithEmailAndPassword({
-    @required String email, @required String password});
+    required String email, required String password});
 
   Stream<AuthUser> get user;
-  Future<AuthUser> getCurrentUser();
 
-  Future<Result<AuthUser>> signInWithEmailAndPassword({final String email,
-    final String password});
+  Future<Result<AuthUser>> signInWithEmailAndPassword({
+    required final String email,
+    required final String password});
   Future<Result<void>> signOut();
   Future<Result<void>> sendPasswordResetEmail(String email);
   Future<Result<AuthUser>> signInAnonymously();
   Future<Result<AuthUser>> signInWithGoogle();
-  Future<Result<AuthUser>> convertUserWithEmail({@required final String email,
-    @required final String password});
+  Future<Result<AuthUser>> convertUserWithEmail({required final String email,
+    required final String password});
   FirebaseAuth getFbAuth();
 }
 
@@ -49,44 +48,39 @@ class AuthServiceImpl implements AuthService{
   final GoogleSignIn _googleSignIn = GoogleSignIn();
 
 
-  AuthServiceImpl({@required this.auth});
+  AuthServiceImpl({required this.auth});
 
   FirebaseAuth getFbAuth() => auth;
 
-  AuthUser _mapUser(User user) {
-    return user != null
-        ? AuthUser.fromFirestore(user)
-        : null ;
+  AuthUser _mapUser(User? user) {
+    return AuthUser(
+      id: user?.uid ,
+        email: user?.email
+    );
   }
 
 
   @override
   Future<Result<AuthUser>> registerWithEmailAndPassword({
-    @required final String email,
-    @required final String password }) async {
+    required final String email,
+    required final String password }) async {
     try{
 
       assert(StringUtils.instance.isNotBlank(email));
       assert(StringUtils.instance.isNotBlank(password));
 
-
-
       UserCredential credentialResult = await auth
           .createUserWithEmailAndPassword(
-          email: email,
-          password: password
-      );
+            email: email,
+            password: password
+          );
 
-
-      if(credentialResult == null)
+      if(credentialResult.user == null)
         return Result.failure(msg: errNull);
 
-      AuthUser user = _mapUser(credentialResult.user);
+      AuthUser user = _mapUser(credentialResult.user!);
 
-      return user == null
-          ? Result.failure(msg: errNull)
-          : Result.success(obj: user);
-
+      return Result.success(obj: user);
 
     } catch (e, stacktrace) {
       return ErrorHandler().handleError(
@@ -102,15 +96,21 @@ class AuthServiceImpl implements AuthService{
   }
 
   @override
-  Future<Result<AuthUser>> convertUserWithEmail({@required final String email,
-    @required final String password}) async{
+  Future<Result<AuthUser>> convertUserWithEmail({required final String email,
+    required final String password}) async{
     try{
 
-      final User currentUser = auth.currentUser;
+      if(auth.currentUser == null)
+        return Result.failure(msg: 'The auth.currentUser is null');
+
+      final User currentUser = auth.currentUser!;
       final AuthCredential credential = EmailAuthProvider
           .credential(email: email, password: password);
 
-      User user = (await currentUser.linkWithCredential(credential)).user;
+      User? user = (await currentUser.linkWithCredential(credential)).user;
+      if(user == null)
+        return Result.failure(msg: 'A null user was received after attempting'
+            ' to link a uid to an email & a password.');
 
       return Result.success(obj: _mapUser(user));
     }
@@ -129,13 +129,23 @@ class AuthServiceImpl implements AuthService{
   @override
   Future<Result<AuthUser>> signInWithGoogle() async{
     try{
-      final GoogleSignInAccount account = await _googleSignIn.signIn();
+      final GoogleSignInAccount? account = await _googleSignIn.signIn();
+
+      if(account == null)
+        return Result.failure(msg: 'Signing with Google account has resulted'
+            ' in a null value.');
+
       final GoogleSignInAuthentication googleAuth = await account.authentication;
       final AuthCredential credential = GoogleAuthProvider.credential(
         idToken: googleAuth.idToken,
         accessToken: googleAuth.accessToken
       );
-      User user = (await auth.signInWithCredential(credential)).user;
+
+      User? user = (await auth.signInWithCredential(credential)).user;
+      if(user == null)
+        return Result.failure(msg: 'A null user was received after attempting'
+            ' to link a uid to an email & a password.');
+
       return Result.success(obj: _mapUser(user));
     }
     catch(e, stacktrace){
@@ -151,38 +161,7 @@ class AuthServiceImpl implements AuthService{
   }
 
 
-  @override
-  Future<Result<AuthUser>> signInWithEmailAndPassword({@required final String email,
-    @required final String password})
-  async {
-    try {
-      assert(StringUtils.instance.isNotBlank(email));
-      assert(StringUtils.instance.isNotBlank(password));
 
-
-      UserCredential credentialResult = await auth
-          .signInWithEmailAndPassword(email: email, password: password);
-
-      if(credentialResult == null || credentialResult.user == null)
-        return Result.failure();
-
-      AuthUser user = AuthUser.fromFirestore(credentialResult.user);
-
-      return user != null
-          ? Result.success(obj: user)
-          : Result.failure();
-
-    } catch (e, stacktrace) {
-      return ErrorHandler().handleError(
-        result: Result.failure(
-          log: Log(
-            stacktrace: stacktrace
-          )
-        ),
-        error: e
-      );
-    }
-  }
 
 
 
@@ -204,11 +183,6 @@ class AuthServiceImpl implements AuthService{
     }
   }
 
-  @override
-  Future<AuthUser> getCurrentUser() {
-    // TODO: implement getCurrentUser
-    return null;
-  }
 
   @override
   Stream<AuthUser> get user {
@@ -235,12 +209,50 @@ class AuthServiceImpl implements AuthService{
     try {
       UserCredential signInResult = await auth.signInAnonymously();
 
-      if(signInResult == null || signInResult.user == null)
+      if(signInResult.user == null)
         return Result.failure();
 
-      AuthUser user = AuthUser.fromFirestore(signInResult.user);
+      AuthUser user = AuthUser(
+        id: signInResult.user!.uid, email: signInResult.user!.email
+      );
 
-      return user != null ? Result.success(obj: user) : Result.failure();
+      return Result.success(obj: user);
+
+    } catch (e, stacktrace) {
+      return ErrorHandler().handleError(
+          result: Result.failure(
+              log: Log(
+                  stacktrace: stacktrace
+              )
+          ),
+          error: e
+      );
+    }
+  }
+
+  @override
+  Future<Result<AuthUser>> signInWithEmailAndPassword({
+    required final String email,
+    required final String password
+  })
+  async {
+    try {
+      assert(StringUtils.instance.isNotBlank(email));
+      assert(StringUtils.instance.isNotBlank(password));
+
+
+      UserCredential credentialResult = await auth
+          .signInWithEmailAndPassword(email: email, password: password);
+
+      if(credentialResult.user == null)
+        return Result.failure();
+
+      AuthUser user = AuthUser(
+          id: credentialResult.user!.uid,
+          email: credentialResult.user!.email
+      );
+
+      return Result.success(obj: user);
 
     } catch (e, stacktrace) {
       return ErrorHandler().handleError(
